requirements.txt
asgiref==3.8.1
attrs==23.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.7
django-js-asset==2.2.0
django-taggit==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.11.0
tzdata==2023.4
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0

pip==24.0
wheel==0.43.0
setuptools==69.2.0